Refactors static methods from Reservation into a service
