+++ app/models/reservation.rb
@@ -17,20 +17,6 @@ class Reservation < ApplicationRecord
   validate :does_not_exceed_office_capacity
   validate :does_not_exceed_workspace_capacity
 
-  def self.reservations_per_day
-    Reservation.group(:date).count(:date)
-  end
-
-  def self.spots_remaining_for_today
-    used_spots = Reservation.where(date: Date.today).count
-    capacity = ReservationPolicy.current.capacity
-    capacity - used_spots
-  end
-
-  def self.for_date(date)
-    Reservation.where(date: date)
-  end
-
   def verified?
     !verified_by.nil?
   end
